support auto-run field
