+++ docker/backend/requirements.txt
@@ -6,7 +6,7 @@ python-dateutil==2.8.1
 # test and development:
 coverage==5.0.4
 pytest==4.3.1
-watchdog==0.9.0
+watchdog==2.2.0
 
 google-auth==1.15.0
 kubernetes==11.0.0
