+++ plynx/db/node.py
@@ -1,26 +1,26 @@
 """Node DB Object and utils"""
-from collections import defaultdict, deque
 from dataclasses import dataclass, field
-from typing import Any, Callable, Dict, List, Optional, Union
+from typing import Any, Callable, List, Optional, Union
 
 from dataclasses_json import dataclass_json
 from past.builtins import basestring
 
-from plynx.constants import Collections, NodeClonePolicy, NodeOrigin, NodeRunningStatus, NodeStatus, ParameterTypes, SpecialNodeId
+from plynx.constants import Collections, NodeClonePolicy, NodeOrigin, NodeRunningStatus, NodeStatus, ParameterTypes
 from plynx.db.db_object import DBObject
 from plynx.plugins.resources.common import FILE_KIND
 from plynx.utils.common import ObjectId
 
 
 # pylint: disable=too-many-branches
-def _clone_update_in_place(node: "Node", node_clone_policy: int, override_finished_state: bool):
+def _clone_update_in_place(node: "Node", node_clone_policy: int, override_finished_state: bool, override_node_id: bool = False):
     if node.node_running_status == NodeRunningStatus.SPECIAL:
         for output in node.outputs:
             output.values = []
         return node
     old_node_id = node._id
     node.template_node_id = node._id
-    node._id = ObjectId()
+    if override_node_id:
+        node._id = ObjectId()
     node.successor_node_id = None
 
     if node_clone_policy == NodeClonePolicy.NODE_TO_NODE:
@@ -208,7 +208,7 @@ class Node(DBObject):
 
     def clone(self, node_clone_policy: int, override_finished_state: bool = True) -> "Node":
         """Return a cloned copy of a Node"""
-        node = _clone_update_in_place(Node.from_dict(self.to_dict()), node_clone_policy, override_finished_state)
+        node = _clone_update_in_place(Node.from_dict(self.to_dict()), node_clone_policy, override_finished_state, override_node_id=True)
         return node
 
     def _get_custom_element(
@@ -244,198 +244,12 @@ class Node(DBObject):
         """Find Log object"""
         return self._get_custom_element(self.logs, name, throw, default=Node._default_log)
 
-    # pylint: disable=inconsistent-return-statements
-    def arrange_auto_layout(self, readonly: bool = False):
-        """Use heuristic to rearange nodes."""
-        # pylint: disable=invalid-name,too-many-locals,too-many-statements
-        HEADER_HEIGHT = 23
-        TITLE_HEIGHT = 20
-        FOOTER_HEIGHT = 10
-        BORDERS_HEIGHT = 2
-        # ITEM_HEIGHT = 20
-        ITEM_HEIGHT = 30
-        OUTPUT_ITEM_HEIGHT = 100
-        SPACE_HEIGHT = 50
-        LEFT_PADDING = 30
-        TOP_PADDING = 80
-        LEVEL_WIDTH = 252
-        SPECIAL_PARAMETER_HEIGHT = 20
-        SPECIAL_PARAMETER_TYPES = [ParameterTypes.CODE]
-        min_node_height = HEADER_HEIGHT + TITLE_HEIGHT + FOOTER_HEIGHT + BORDERS_HEIGHT
-
-        node_id_to_level = defaultdict(lambda: -1)
-        node_id_to_node = {}
-        queued_node_ids = set()
-        children_ids = defaultdict(set)
-
-        sub_nodes = self.get_parameter_by_name('_nodes').value.value
-
-        if len(sub_nodes) == 0:
-            return
-
-        node_ids = {node._id for node in sub_nodes}
-        non_zero_node_ids = set()
-        for node in sub_nodes:
-            node_id_to_node[node._id] = node
-            for input in node.inputs:   # pylint: disable=redefined-builtin
-                for input_reference in input.input_references:
-                    parent_node_id = ObjectId(input_reference.node_id)
-                    non_zero_node_ids.add(parent_node_id)
-                    children_ids[parent_node_id].add(node._id)
-
-        leaves = node_ids - non_zero_node_ids
-        to_visit: deque = deque()
-        # Alwasy put Output Node in the end
-        push_special = SpecialNodeId.OUTPUT in leaves and len(leaves) > 1
-        for leaf_id in leaves:
-            node_id_to_level[leaf_id] = 1 if push_special and leaf_id != SpecialNodeId.OUTPUT else 0
-            to_visit.append(leaf_id)
-
-        while to_visit:
-            node_id = to_visit.popleft()
-            node = node_id_to_node[node_id]
-            node_level = max([node_id_to_level[node_id]] + [node_id_to_level[child_id] + 1 for child_id in children_ids[node_id]])
-            node_id_to_level[node_id] = node_level
-            for input in node.inputs:   # pylint: disable=redefined-builtin
-                for input_reference in input.input_references:
-                    parent_node_id = ObjectId(input_reference.node_id)
-                    parent_level = node_id_to_level[parent_node_id]
-                    node_id_to_level[parent_node_id] = max(node_level + 1, parent_level)
-                    if parent_node_id not in queued_node_ids:
-                        to_visit.append(parent_node_id)
-                        queued_node_ids.add(parent_node_id)
-
-        max_level = max(node_id_to_level.values())
-        level_to_node_ids: Dict[int, List[ObjectId]] = defaultdict(list)
-        row_heights: Dict[int, int] = defaultdict(lambda: 0)
-
-        def get_index_helper(node, level):
-            if level < 0:
-                return 0
-            parent_node_ids = set()
-            for input in node.inputs:   # pylint: disable=redefined-builtin
-                for input_reference in input.input_references:
-                    parent_node_ids.add(ObjectId(input_reference.node_id))
-
-            for index, node_id in enumerate(level_to_node_ids[level]):
-                if node_id in parent_node_ids:
-                    return index
-            return -1
-
-        def get_index(node, max_level, level):
-            # pylint: disable=consider-using-generator
-            return tuple(
-                [get_index_helper(node, lvl) for lvl in range(max_level, level, -1)]
-            )
-
-        for node_id, level in node_id_to_level.items():
-            level_to_node_ids[level].append(node_id)
-
-        # Push Input Node up the level
-        if SpecialNodeId.INPUT in node_id_to_level and \
-                (node_id_to_level[SpecialNodeId.INPUT] != max_level or len(level_to_node_ids[max_level]) > 1):
-            input_level = node_id_to_level[SpecialNodeId.INPUT]
-            level_to_node_ids[input_level] = [node_id for node_id in level_to_node_ids[input_level] if node_id != SpecialNodeId.INPUT]
-            max_level += 1
-            node_id_to_level[SpecialNodeId.INPUT] = max_level
-            level_to_node_ids[max_level] = [SpecialNodeId.INPUT]
-
-        for level in range(max_level, -1, -1):
-            level_node_ids = level_to_node_ids[level]
-            index_to_node_id = []
-            for node_id in level_node_ids:
-                node = node_id_to_node[node_id]
-                index = get_index(node, max_level, level)
-                index_to_node_id.append((index, node_id))
-
-            index_to_node_id.sort()
-            level_to_node_ids[level] = [node_id for _, node_id in index_to_node_id]
-
-            for index, node_id in enumerate(level_to_node_ids[level]):
-                node = node_id_to_node[node_id]
-                special_parameters_count = sum(
-                    1 if parameter.parameter_type in SPECIAL_PARAMETER_TYPES and parameter.widget else 0
-                    for parameter in node.parameters
-                )
-                node_height = sum([
-                    min_node_height,
-                    ITEM_HEIGHT * len(node.inputs) + OUTPUT_ITEM_HEIGHT * len(node.outputs),
-                    special_parameters_count * SPECIAL_PARAMETER_HEIGHT
-                ])
-                row_heights[index] = max(row_heights[index], node_height)
-
-        # TODO compute grid in a separate function
-        if readonly:
-            return level_to_node_ids, node_id_to_node
-
-        cum_heights = [0]
-        for row_height in row_heights:
-            cum_heights.append(cum_heights[-1] + row_height + SPACE_HEIGHT)
-
-        max_height = max(cum_heights)
-
-        for level in range(max_level, -1, -1):
-            level_node_ids = level_to_node_ids[level]
-            level_height = cum_heights[len(level_node_ids)]
-            level_padding = (max_height - level_height) // 2
-            for index, node_id in enumerate(level_node_ids):
-                node = node_id_to_node[node_id]
-                node.x = LEFT_PADDING + (max_level - level) * LEVEL_WIDTH
-                node.y = TOP_PADDING + level_padding + cum_heights[index]
-
-    def reset_failed_nodes(self):
-        """
-        Reset the statuses of the failed subnodes.
-
-        Resetting means node_running_status as well as the outputs and the logs.
-        """
+    def get_sub_nodes(self):
+        """Get a list of subnodes"""
         sub_nodes_parameter = self.get_parameter_by_name('_nodes', throw=False)
         if not sub_nodes_parameter:
-            return
-
-        sub_nodes = sub_nodes_parameter.value.value
-        for sub_node in sub_nodes:
-            if not NodeRunningStatus.is_failed(sub_node.node_running_status):
-                continue
-            sub_node.node_running_status = NodeRunningStatus.READY
-            for resource in sub_node.outputs + sub_node.logs:
-                resource.values = []
-
-    def apply_cache(self):
-        """Apply cache values to outputs and logs"""
-        sub_nodes_parameter = self.get_parameter_by_name('_nodes', throw=False)
-        if not sub_nodes_parameter:
-            # TODO check if cacheable
-            raise NotImplementedError("Subnodes not found. Do we want to be here?")
-
-        sub_nodes = sub_nodes_parameter.value.value
-
-        for sub_node in sub_nodes:
-            if not sub_node._cached_node:
-                continue
-            sub_node.node_running_status = sub_node._cached_node.node_running_status
-            sub_node.outputs = sub_node._cached_node.outputs
-            sub_node.logs = sub_node._cached_node.logs
-
-            sub_node._cached_node = None
-
-    def is_all_cached_and_successful(self):
-        """
-        Check if the Node in a run needs recomputing at all.
-        """
-        sub_nodes_parameter = self.get_parameter_by_name('_nodes', throw=False)
-        if not sub_nodes_parameter:
-            # TODO check if cacheable
-            return True
-
-        sub_nodes = sub_nodes_parameter.value.value
-
-        for sub_node in sub_nodes:
-            if not sub_node._cached_node:
-                return False
-            if NodeRunningStatus.is_failed(sub_node._cached_node.node_running_status):
-                return False
-        return True
+            raise Exception("Subnodes not found")
+        return sub_nodes_parameter.value.value
 
 
 @dataclass_json
